# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Wegweiser API
  version: 0.1.0
  description: |
    The API every Wegweiser client speaks. The web interface and the `weg`
    command line are both clients of it, and neither of them reaches the
    database any other way (architecture invariant 1).

    This document is the source of truth. Clients are generated from it, so an
    endpoint that is not written down here does not exist.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: /api/v1
security:
  - bearerToken: []
...
paths:
...
  /queries/stream:
    get:
      operationId: streamQueries
      tags: [queries]
      summary: Watch queries as they are answered
      description: |
        A live tail of the exchanges matching the filter, as Server-Sent
        Events. Two kinds of message arrive:

        - `query`, one per exchange, carrying a `QueryEvent`.
        - `status`, whenever what the stream is leaving out changes and every
          few seconds regardless, carrying a `StreamStatus`. It says what the
          stream is sampling at and how much it has dropped, because a stream
          that quietly leaves things out while looking complete is worse than
          one that admits to showing every fiftieth query
          (docs/decisions/ D9).

        The filter is applied before anything is buffered, so watching one zone
        stays complete however busy the rest of the server is. A watcher that
        stops reading loses its oldest exchanges, never its newest: this is a
        live view, and it is never allowed to slow a query down.

        How many watchers may be open at once is bounded, because each one
        costs the query path a filter evaluation per query. A request past the
        bound is refused with 503.
      parameters:
        - name: name
          in: query
          description: |
            Watch a name and everything below it. Matching is
            case-insensitive (RFC 4343), so it finds a query however 0x20
            encoding wrote it.
          schema: { type: string }
        - name: type
          in: query
          description: Watch only these question types. Repeat for more than one.
          schema:
            type: array
            items: { type: string }
        - name: rcode
          in: query
          description: |
            Watch only these response codes, by mnemonic, `NXDOMAIN`,
            `REFUSED`. Repeat for more than one.
          schema:
            type: array
            items: { type: string }
        - name: client
          in: query
          description: |
            Watch only this address or network, in CIDR notation. A bare
            address is taken as a single host.
          schema: { type: string }
      responses:
        "200":
          description: The stream, until the client goes away.
          content:
            text/event-stream:
              schema: { type: string }
        default: { $ref: "#/components/responses/Problem" }
...
  /auth/session:
    post:
      operationId: createSession
      tags: [auth]
      summary: Exchange a token for a browser session
      description: |
        The web interface does not hold a long-lived token. It posts one here
        once and receives an `httpOnly` session cookie, which no script that
        gets injected into the page can read, which is the whole point.

        A second, readable cookie carries a CSRF value. Every state-changing
        request authenticated by the session cookie has to repeat it in the
        `X-Wegweiser-CSRF` header. `SameSite=Strict` already blocks a
        cross-site form post; this is the second layer. A request that
        authenticates with a bearer token skips the check, because CSRF needs
        ambient credentials and a bearer token is not ambient.

        Unauthenticated, because a request that could authenticate would not
        need this. The token inside is checked the same way and is rate-limited
        the same way.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSession" }
      responses:
        "201":
          description: The session was opened. The cookies are in the headers.
          headers:
            Set-Cookie:
              description: The session cookie and the CSRF cookie.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        default: { $ref: "#/components/responses/Problem" }

    get:
      operationId: getSession
      tags: [auth]
      summary: Report who this request is authenticated as
      description: |
        What an interface asks when it loads, to find out whether it has to
        show a login at all, and what to grey out if it does not.
      responses:
        "200":
          description: The caller.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        default: { $ref: "#/components/responses/Problem" }

    delete:
      operationId: deleteSession
      tags: [auth]
      summary: End the session
      description: |
        The cookies are cleared and the session is forgotten here, so a copy of
        the cookie taken beforehand stops working too. The token it was opened
        with is untouched.
      responses:
        "204":
          description: The session is over.
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: |
        A named API token. Tokens carry the scopes `read`, `write` and `admin`;
        an endpoint that changes anything needs `write`.
...
  responses:
    Problem:
      description: The request could not be carried out.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: |
        A failure, in the shape of RFC 9457. Anything a client has to act on
        travels as a field rather than as prose in `detail`, because no client
        should be parsing English.
      required: [type, title, status]
      properties:
        type:
          type: string
          description: A URI reference naming the kind of failure.
        title:
          type: string
          description: A short, stable summary of the kind of failure.
        status:
          type: integer
          description: The HTTP status code, repeated for clients that lost it.
        detail:
          type: string
          description: What went wrong with this particular request.
        instance:
          type: string
          description: The request this failure belongs to.
        conflicts:
          type: array
          description: Reverse entries that were not written because another name already claims the address.
          items: { $ref: "#/components/schemas/Conflict" }
        missingZones:
          type: array
          description: Reverse zones that would be needed and do not exist.
          items: { $ref: "#/components/schemas/MissingZone" }
    ReverseConflictPolicy:
      type: string
      enum: [first-wins, last-wins, multi, reject]
      description: |
        What happens when an address already answers a reverse lookup with some
        other name (docs/decisions/ D3).

        - `first-wins` keeps the entry that is there and reports the conflict.
          The default, and the only value that never changes an answer nobody
          asked to change.
        - `last-wins` replaces a generated entry with the new one, still
          reporting the conflict. An entry somebody wrote by hand is never
          replaced.
        - `multi` keeps both, which is the literal reading of "generate the
          reverse entry" and turns a routine change into a multi-record PTR set
          that reverse-lookup checks are not built for.
        - `reject` refuses the whole change.
...
    CreateSession:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: The API token to open the session with.
    Session:
      type: object
      description: Who a request is authenticated as, and what it may do.
      required: [name, scopes]
      properties:
        name:
          type: string
          description: The name of the token this authenticates as.
        scopes:
          type: array
          items: { $ref: "#/components/schemas/Scope" }
        expiresAt:
          type: string
          format: date-time
          description: When the session ends. Absent when the request authenticated with a bearer token.
    Scope:
      type: string
      description: |
        What a token may do. The scopes are ordered (admin allows everything
        write allows, and write everything read allows) so a token carries the
        strongest one it needs rather than a list.
      enum: [read, write, admin]
...
    Conflict:
      type: object
      description: |
        An address already answers with another name (docs/decisions/ D3). It
        is data rather than an error: which name should be canonical is a
        decision for a person.
      required: [address, existingName, requestedName, policy]
      properties:
        address: { type: string }
        existingName: { type: string }
        requestedName: { type: string }
        existingRecordId: { type: string }
        policy:
          allOf:
            - $ref: "#/components/schemas/ReverseConflictPolicy"
          description: |
            The policy that decided what happened here, so a client can say
            whether the existing name was kept, replaced or joined rather than
            leaving the reader to guess from the records.
    MissingZone:
      type: object
      description: |
        A reverse entry has nowhere to go, because no zone covers the address.
        Creating one is an assertion of authority, so it is offered rather than
        done (docs/decisions/ D6).
      required: [address, zoneName]
      properties:
        address: { type: string }
        zoneName:
          type: string
          description: The reverse zone that would be needed.